+++ components/src/main/java/com/mercadolibre/android/andesui/button/hierarchy/AndesButtonHierarchyInterface.kt
@@ -40,10 +40,9 @@ internal sealed class AndesButtonHierarchyInterface {
      * Returns an [Int] representing a @ColorInt that will be used when tinting the icon.
      *
      * @param context needed for accessing some resources.
-     * @return an [Int] representing a @ColorInt that will be used when tinting the icon.
+     * @return a [ColorStateList] that contains the data for the icon color.
      */
-    @ColorInt
-    abstract fun iconColor(context: Context): Int
+    fun iconColor(context: Context) = textColor(context)
 
     /**
      * Returns the [Typeface] that should be used for the text inside the [AndesButton].
@@ -62,7 +61,6 @@ internal sealed class AndesButtonHierarchyInterface {
 internal object AndesLoudButtonHierarchy : AndesButtonHierarchyInterface() {
     override fun background(context: Context, cornerRadius: Float) = getConfiguredBackground(context, cornerRadius, createBackgroundColorConfigLoud())
     override fun textColor(context: Context) = getConfiguredTextColor(context, createTextColorConfigLoud())
-    override fun iconColor(context: Context) = ContextCompat.getColor(context, R.color.andesui_button_loud_text)
 }
 
 /**
@@ -73,7 +71,6 @@ internal object AndesLoudButtonHierarchy : AndesButtonHierarchyInterface() {
 internal object AndesQuietButtonHierarchy : AndesButtonHierarchyInterface() {
     override fun background(context: Context, cornerRadius: Float) = getConfiguredBackground(context, cornerRadius, createBackgroundColorConfigQuiet())
     override fun textColor(context: Context) = getConfiguredTextColor(context, createTextColorConfigQuiet())
-    override fun iconColor(context: Context) = ContextCompat.getColor(context, R.color.andesui_button_quiet_text)
 }
 
 /**
@@ -84,5 +81,4 @@ internal object AndesQuietButtonHierarchy : AndesButtonHierarchyInterface() {
 internal object AndesTransparentButtonHierarchy : AndesButtonHierarchyInterface() {
     override fun background(context: Context, cornerRadius: Float) = getConfiguredBackground(context, cornerRadius, createBackgroundColorConfigTransparent())
     override fun textColor(context: Context) = getConfiguredTextColor(context, createTextColorConfigTransparent())
-    override fun iconColor(context: Context) = ContextCompat.getColor(context, R.color.andesui_button_transparent_text)
 }