+++ CHANGELOG.md
@@ -1,4 +1,7 @@
-# v1.0.1
+# v1.1.0
+# Added
+- Initial release for AndesMessage.
+
 # Fixed
 - Icon wasn't changing when AndesButton was disabled.
 - Changelog updated